SERVER/main_thought_stream/requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aiven-client==4.2.1
apturl==0.5.2
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
awscli==1.34.20
bcrypt==3.2.0
blinker==1.9.0
boto3==1.35.15
botocore==1.35.20
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.6
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
djangorestframework==3.15.2
docutils==0.16
duplicity==0.8.21
executing==2.1.0
fasteners==0.14.1
Flask==3.1.0
frozenlist==1.5.0
future==0.18.2
greenlet==3.1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
meld==3.20.4
monotonic==1.6
more-itertools==8.10.0
multidict==6.1.0
mysqlclient==2.1.1
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==2.1.1
oauthlib==3.2.0
olefile==0.46
packaging==24.1
pandas==2.2.3
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.12.4
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyspark==3.5.3
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
roman==3.3
rsa==4.7.2
s3transfer==0.10.2
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.2
stack-data==0.6.3
systemd-python==234
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
